refactor main into small components
