+++ data/data11-10-02.txt
@@ -1,0 +1,5 @@
+11:10:17; oro has entered the house.$
+11:10:19; oro is sitting on a(n) chair.$
+11:10:29; oro is sitting on a(n) chair.$
+11:10:40; oro is sitting on a(n) chair.$
+11:10:45; oro is sitting on a(n) chair.$
