+++ data/data11-12-16.txt
@@ -1,0 +1,14 @@
+11:12:32; oro has entered the house.$
+11:12:34; oro is sitting on a(n) sofa.$
+11:12:36; oro is sitting on a(n) sofa.$
+11:12:40; oro is sitting on a(n) sofa.$
+11:12:44; oro is sitting on a(n) sofa.$
+11:12:48; oro is sitting on a(n) sofa.$
+11:12:55; oro is sitting on a(n) sofa.$
+11:12:59; oro is sitting on a(n) sofa.$
+11:13:03; oro is sitting on a(n) sofa.$
+11:13:07; oro is sitting on a(n) sofa.$
+11:13:11; oro is sitting on a(n) sofa.$
+11:13:14; oro is sitting on a(n) sofa.$
+11:13:18; oro is sitting on a(n) sofa.$ oro is using her phone.$
+11:13:22; oro is using her phone.$ oro is sitting on a(n) sofa.$
