+++ forms/form11-13-15.txt
@@ -1,0 +1,22 @@
+report time: 11-13-15
+
+Important events:
+
+
+Situations:
+ oro has entered the house from 11:12:32 to 11:12:32.
+ oro is sitting on a(n) sofa from 11:12:34 to 11:13:14.
+
+
+In total:
+ oro has entered the house for 0:00:00 second(s), 1 time(s) total.
+ oro is sitting on a(n) sofa for 0:00:40 second(s), 1 time(s) total.
+ oro has entered the house from 11:12:32 to 11:12:32.
+ oro has entered the house from 11:12:32 to 11:12:32.
+ oro is sitting on a(n) sofa from 11:12:34 to 11:13:14.
+ oro is sitting on a(n) sofa from 11:12:34 to 11:13:14.
+
+
+In total:
+ oro has entered the house for 00:00:00 second(s), 2 time(s) total.
+ oro is sitting on a(n) sofa for 00:01:20 minute(s), 2 time(s) total.
