+++ src/Object.py
@@ -1,0 +1,5 @@
+class Object:
+    def __init__(self, class_id, score, bbox):
+        self.id = class_id
+        self.score = score
+        self.bbox = bbox
