+++ src/deprecated/Camera.py
