+++ src/image_utils.py
@@ -1,0 +1,13 @@
+import cv2
+
+
+def draw_labels(boxes, colors, class_ids, classes, img, indexes):
+    font = cv2.FONT_HERSHEY_PLAIN
+    for i in range(len(boxes)):
+        if i in indexes:
+            x, y, w, h = boxes[i]
+            label = str(classes[class_ids[i]])
+            color = colors[i]
+            cv2.rectangle(img, (x, y), (x + w, y + h), color, 2)
+            cv2.putText(img, label, (x, y - 5), font, 1, color, 1)
+    cv2.imshow("Image", img)
